+++ settings.gradle
@@ -37,3 +37,4 @@ include 'spring-pulsar-sample-apps:sample-reactive'
 include 'spring-pulsar-sample-apps:sample-pulsar-binder'
 include 'spring-pulsar-docs'
 include 'spring-pulsar-spring-cloud-stream-binder'
+include 'spring-pulsar-test'
