+++ spring-pulsar-test/build.gradle
@@ -1,0 +1,18 @@
+plugins {
+    id 'org.springframework.pulsar.spring-module'
+}
+
+description = 'Spring Pulsar Test Module'
+
+dependencies {
+    implementation 'org.testcontainers:pulsar'
+    implementation 'org.testcontainers:junit-jupiter'
+}
+
+project.afterEvaluate {
+    project.tasks.publishArtifacts.enabled(false)
+    project.tasks.artifactoryPublish.enabled(false)
+    project.tasks.publishToOssrh.enabled(false)
+    project.tasks.publishMavenJavaPublicationToOssrhRepository.enabled(false)
+    project.tasks.publishAllPublicationsToOssrhRepository.enabled(false)
+}
