+++ spring-pulsar/src/test/java/org/springframework/pulsar/core/PulsarProducerFactoryTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 the original author or authors.
+ * Copyright 2022-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -37,6 +37,8 @@ import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import org.springframework.pulsar.test.support.PulsarTestContainerSupport;
 
 /**
  * Common tests for {@link DefaultPulsarProducerFactory} and
